db: port changesets, queue summarizer, and all writers to Diesel
The big port: db/changesets.rs's 18 helpers now take &mut SqliteConnection
instead of &Transaction / &Connection. Simple inserts/updates use the
Diesel DSL. The two multi-JOIN read queries
(query_change_set_for_commit_pair, query_change_set_for_base_with_hashes)
stay as raw SQL via diesel::sql_query + QueryableByName per the migration
plan's directive against full DSL ports for gnarly aggregate queries.

Callers updated to acquire a pooled connection and wrap atomic groups in
conn.transaction(|conn| ...):
  - db/store_bare_changeset.rs, store_new_changeset.rs, store_evolved_changeset.rs
  - summarize/queue_summarizer.rs (8 connection-open sites + ~10 changesets
    callers consolidated into Diesel transactions). The worker now holds a
    cloned DbPool instead of a PathBuf.
  - summarize/find_existing.rs drops the db_path parameter from
    for_current_state and by_base_with_hashes; pool reads are enough.
  - state/build_state.rs's record_build and matches_current_state now
    drive the changesets reader through the pool.
  - summarize/pipelines/fresh_changeset.rs and evolved_changeset.rs drop
    their db_path parameter — propagated to all transitive callers
    (summarize::new_changeset, summarize::pipelines::history).
  - managed_edits/managed_edit.rs drops the now-unused ManagedEditContext.db_path.

Tables.rs: line_count and attempted_count promoted to BigInt to match
the i64 fields on Change and QueuedSummary.

+++ apps/native/src-tauri/src/db/tables.rs
@@ -35,7 +35,7 @@ diesel::table! {
         hash -> Text,
         filename -> Text,
         diff -> Text,
-        line_count -> Integer,
+        line_count -> BigInt,
         created_at -> BigInt,
         own_summary_id -> Nullable<BigInt>,
     }
@@ -71,7 +71,7 @@ diesel::table! {
     queued_summaries (id) {
         id -> BigInt,
         status -> Text,
-        attempted_count -> Integer,
+        attempted_count -> BigInt,
         prompt -> Text,
         model_response -> Nullable<Text>,
         group_summary_id -> Nullable<BigInt>,
